🎨 Improve WebSocket & Game structure
