+++ src/router/index.js
@@ -2,13 +2,15 @@
   * @Author: hiddenSharp429 z404878860@163.com
   * @Date: 2024-10-28 19:51:51
  * @LastEditors: hiddenSharp429 z404878860@163.com
- * @LastEditTime: 2024-11-14 12:43:28
+ * @LastEditTime: 2024-11-14 23:22:14
   */
   import { createRouter, createWebHistory } from 'vue-router'
   import Home from '@/components/Home.vue'
+  import Auth from '@/components/Auth.vue'
   import Layout from '@/layouts/Layout.vue'
   import Settings from '@/views/Settings.vue'
-  import GameLobby from '../views/GameLobby.vue'
+  import GameLobby from '@/views/GameLobby.vue'
+  import GameRoom from '@/views/GameRoom.vue'
 
   const routes = [
     {
@@ -23,13 +25,12 @@
         {
           path: 'auth',
           name: 'auth',
-          component: () => import('../components/Auth.vue')
+          component: Auth
         },
-        // 添加测试路由
         {
-          path: 'test-room/:id?',
-          name: 'testRoom',
-          component: () => import('../components/GameRoom.vue')
+          path: '/room/:id?',
+          name: 'room',
+          component: GameRoom
         },
         {
           path: 'settings',
